Extract creation of context menu
