+++ app/src/main-process/main.ts
@@ -22,6 +22,7 @@ import {
 import { now } from './now'
 import { showUncaughtException } from './show-uncaught-exception'
 import { IMenuItem } from '../lib/menu-item'
+import { buildContextMenu } from './menu/build-context-menu'
 
 enableSourceMaps()
 
@@ -274,41 +275,9 @@ app.on('ready', () => {
   ipcMain.on(
     'show-contextual-menu',
     (event: Electron.IpcMessageEvent, items: ReadonlyArray<IMenuItem>) => {
-      const menu = new Menu()
-
-      for (const [ix, item] of items.entries()) {
-        // Special case editMenu in context menus. What we
-        // mean by this is that we want to insert all edit
-        // related menu items into the menu at this spot, we
-        // don't want a sub menu
-        if (item.role === 'editMenu') {
-          const editMenu = Menu.buildFromTemplate([item]).items[0]
-
-          if (!editMenu.submenu) {
-            continue
-          }
-          // We don't use styled inputs anywhere at the moment
-          // so let's skip this for now and when/if we do we
-          // can make it configurable from the callee
-          editMenu.submenu.items
-            .filter(
-              editItem =>
-                editItem.role &&
-                editItem.role.toLocaleLowerCase() !== 'pasteandmatchstyle'
-            )
-            .forEach(editItem => menu.append(editItem))
-        }
-
-        menu.append(
-          new MenuItem({
-            label: item.label,
-            type: item.type,
-            enabled: item.enabled,
-            role: item.role,
-            click: () => event.sender.send('contextual-menu-action', ix),
-          })
-        )
-      }
+      const menu = buildContextMenu(items, item => {
+        event.sender.send('contextual-menu-action', items.indexOf(item))
+      })
 
       const window = BrowserWindow.fromWebContents(event.sender)
       menu.popup(window, { async: true })
